src/main/java/hello/hellospring/repsitory/MemoryMemberRepository.java
package hello.hellospring.repsitory;

import hello.hellospring.domain.Member;

import java.util.*;

public class MemoryMemberRepository implements MemberRepository {

    private static Map<Long, Member> store = new HashMap<>();   //동시성 문제를 제거하기 위한 다른 클래스(?)를 사용해야 한다.
    private static long sequence = 0L;  //동시성 문제를 제거하기 위한 다른 클래스(?)를 사용해야 한다.

    @Override
    public Member save(Member member) {
        member.setId((++sequence));
        store.put(member.getId(), member);
        return member;
    }

    @Override
    public Optional<Member> findById(Long id) {
        //Optional.ofNullable --> null인 경우를 처리하기 위함.
        return Optional.ofNullable(store.get(id));
    }

    @Override
    public Optional<Member> findByName(String name) {
       //람다식을 사용
        return store.values().stream()
               .filter(member -> member.getName().equals(name))
               .findAny();
    }

    @Override
    public List<Member> findAll() {
        return new ArrayList<>(store.values());
    }

    public void clearStore() {
        store.clear();
    }
}
